+++ src/SlideDotNet/Services/Drawing/ISlideSchemeService.cs
@@ -1,9 +1,12 @@
-﻿using SlideDotNet.Models;
+﻿using System.IO;
+using SlideDotNet.Models;
 
 namespace SlideDotNet.Services.Drawing
 {
     public interface ISlideSchemeService
     {
-        void SaveScheme(string filePath, ShapeCollection shapesValue, int sldW, int sldH);
+        void SaveScheme(ShapeCollection shapes, int sldW, int sldH, string filePath);
+        
+        void SaveScheme(ShapeCollection shapes, int sldW, int sldH, Stream stream);
     }
 }